Extract code from `no-array-reverse`
